Move to esbuild
