+++ .npmrc
@@ -1,0 +1,2 @@
+public-hoist-pattern[]=*@one-ini/wasm
+public-hoist-pattern[]=*@cdktf/hcl2json
